Clean Code

+++ Forms/F_Instrutor.cs
@@ -31,7 +31,7 @@ namespace Academia_AMS
             Personal persona = new Personal(); // Criando uma instância de Personal
             persona.T_NAME = txt_Nome.Text;
             persona.T_SERVICO = comboBox1.Text;
-            persona.salario = double.Parse(txt_Salario.Text); // Alterando para double, pois parece ser um valor decimal
+            persona.salario = long.Parse(txt_Salario.Text); // Alterando para double, pois parece ser um valor decimal
             persona.contrato = txt_Contrato.Text;
             persona.N_CPF = int.Parse(txt_Cpf.Text);
             persona.T_OBS = txt_Obs.Text;
